fix: (글 수정) 저장된 미디어 복붙 시 유지 처리, 삭제 미디어 목록 중복 제거
## 작업 내용
- 글 수정 시, 이미 저장된 이미지/동영상을 에디터 안에서 복사·붙여넣기 해도
  "기존 미디어 유지"로 처리되도록 수정했습니다.
  (복붙한 미디어가 새 업로드로 잘못 치거나 placeholder 치환에 포함되던 문제 해결)

- 수정 시 "삭제된 미디어" 목록에 같은 미디어가 여러 번 들어가던 현상을 수정했습니다.

- 위 동작을 위해 에디터 붙여넣기·미디어 치환 로직을 정리했습니다.
