+++ components/tiptap/tiptap-node/video-node/video-extension.ts
@@ -33,12 +33,22 @@ export const Video = Node.create<VideoOptions>({
       controls: {
         default: true,
       },
+      // 처음 수정되는 대상에 data-media-id 추가 -> new 미디어와 구분짓기 위함
       dataMediaId: {
         default: null,
         parseHTML: (element) => element.getAttribute('data-media-id'),
         renderHTML: (attributes) =>
           attributes.dataMediaId
             ? { 'data-media-id': attributes.dataMediaId }
+            : {},
+      },
+      // replaceVideosWithPlaceholders에서 이 속성을 보고 치환 제외하기 위함
+      dataSkipPlaceholder: {
+        default: null,
+        parseHTML: (element) => element.getAttribute('data-skip-placeholder'),
+        renderHTML: (attributes) =>
+          attributes.dataSkipPlaceholder
+            ? { 'data-skip-placeholder': attributes.dataSkipPlaceholder }
             : {},
       },
     };
@@ -53,7 +63,10 @@ export const Video = Node.create<VideoOptions>({
   },
 
   renderHTML({ HTMLAttributes }) {
-    return ['video', mergeAttributes(this.options.HTMLAttributes, HTMLAttributes)];
+    return [
+      'video',
+      mergeAttributes(this.options.HTMLAttributes, HTMLAttributes),
+    ];
   },
 
   addCommands() {
